Strip IR decode
